requirements.txt
CairoSVG==2.7.1
customtkinter==5.2.2
matplotlib==3.7.2
numpy==1.24.3
openpyxl==3.0.9
pandas==2.0.3
Pillow==11.1.0
scipy==1.10.1
